feat: rename package to datascience-agent with dsagent import
- PyPI package name: datascience-agent
- Import name: dsagent (from dsagent import PlannerAgent)
- CLI command: dsagent
- Environment variables: DSAGENT_*

This provides a short, memorable import name while using the
descriptive datascience-agent name on PyPI.
