+++ .env.example
@@ -1,5 +1,5 @@
 # =============================================================================
-# DataPilot - Configuration
+# DSAgent - Configuration
 # =============================================================================
 # Copy this file to .env and fill in your values
 # Priority: CLI args > Environment variables > .env file > defaults
@@ -54,16 +54,16 @@ GOOGLE_API_KEY=your-google-api-key-here
 # =============================================================================
 
 # Maximum rounds before stopping (default: 30)
-# DATAPILOT_MAX_ROUNDS=30
+# DSAGENT_MAX_ROUNDS=30
 
 # LLM temperature (default: 0.3)
-# DATAPILOT_TEMPERATURE=0.3
+# DSAGENT_TEMPERATURE=0.3
 
 # Max tokens per response (default: 4096)
-# DATAPILOT_MAX_TOKENS=4096
+# DSAGENT_MAX_TOKENS=4096
 
 # Code execution timeout in seconds (default: 300)
-# DATAPILOT_CODE_TIMEOUT=300
+# DSAGENT_CODE_TIMEOUT=300
 
 # Default workspace directory (default: ./workspace)
-# DATAPILOT_WORKSPACE=./workspace
+# DSAGENT_WORKSPACE=./workspace
