+++ src/dsagent/__init__.py
@@ -8,14 +8,14 @@ This package provides a powerful agent for data analysis tasks that:
 - Streams events for real-time UI updates
 
 Basic usage:
-    from datapilot import PlannerAgent
+    from dsagent import PlannerAgent
 
     with PlannerAgent(model="gpt-4o") as agent:
         result = agent.run("Analyze sales data and find trends")
         print(result.answer)
 
 With streaming:
-    from datapilot import PlannerAgent, EventType
+    from dsagent import PlannerAgent, EventType
 
     agent = PlannerAgent()
     for event in agent.run_stream("Build a model"):
@@ -23,8 +23,8 @@ With streaming:
             print(f"Code: {event.success}")
 """
 
-from datapilot.agents.base import PlannerAgent, AgentResult
-from datapilot.schema.models import (
+from dsagent.agents.base import PlannerAgent, AgentResult
+from dsagent.schema.models import (
     AgentConfig,
     AgentEvent,
     EventType,
@@ -33,13 +33,13 @@ from datapilot.schema.models import (
     PlanStep,
     SessionState,
 )
-from datapilot.core.executor import JupyterExecutor
-from datapilot.core.engine import AgentEngine
-from datapilot.core.planner import PlanParser
-from datapilot.core.context import RunContext
-from datapilot.utils.logger import AgentLogger, Colors
-from datapilot.utils.notebook import NotebookBuilder, ExecutionTracker
-from datapilot.utils.run_logger import RunLogger
+from dsagent.core.executor import JupyterExecutor
+from dsagent.core.engine import AgentEngine
+from dsagent.core.planner import PlanParser
+from dsagent.core.context import RunContext
+from dsagent.utils.logger import AgentLogger, Colors
+from dsagent.utils.notebook import NotebookBuilder, ExecutionTracker
+from dsagent.utils.run_logger import RunLogger
 
 __version__ = "0.3.0"
