+++ src/dsagent/agents/__init__.py
@@ -1,6 +1,6 @@
 """Agent implementations for the AI Planner."""
 
-from datapilot.agents.base import PlannerAgent
+from dsagent.agents.base import PlannerAgent
 
 __all__ = [
     "PlannerAgent",
