+++ src/dsagent/core/__init__.py
@@ -1,0 +1,11 @@
+"""Core modules for the AI Planner Agent."""
+
+from dsagent.core.executor import JupyterExecutor
+from dsagent.core.planner import PlanParser
+from dsagent.core.engine import AgentEngine
+
+__all__ = [
+    "JupyterExecutor",
+    "PlanParser",
+    "AgentEngine",
+]
