+++ src/dsagent/utils/__init__.py
@@ -1,7 +1,7 @@
 """Utility modules for the AI Planner Agent."""
 
-from datapilot.utils.logger import AgentLogger
-from datapilot.utils.notebook import NotebookBuilder, ExecutionTracker
+from dsagent.utils.logger import AgentLogger
+from dsagent.utils.notebook import NotebookBuilder, ExecutionTracker
 
 __all__ = [
     "AgentLogger",
